exponent in 64 bits

+++ include/io/CliParser.hpp
@@ -9,7 +9,7 @@
 namespace io {
 
 struct CliOptions {
-    uint32_t exponent = 0;
+    uint64_t exponent = 0;
     uint32_t iterforce = 0;
     int device_id = 0;
     bool tune = false;
